backend/requirements.txt
asgiref==3.9.1
distlib==0.4.0
Django==5.2.6
django-cors-headers==4.8.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
pillow==11.3.0
platformdirs==4.4.0
PyJWT==2.10.1
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.34.0
wheel==0.45.1
